add function getColumns to create columns dynamically
