+++ src/features/duplicates/components/NotDuplicatesModalList.tsx
@@ -1,16 +1,14 @@
 import { FC } from 'react';
-import { Box, Button, FormControl, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import { DataGridPro, GridColDef } from '@mui/x-data-grid-pro';
 
+import getColumns from '../utils/getColumns';
 import messageIds from '../l10n/messageIds';
 import { PotentialDuplicate } from '../store';
 import useDuplicatesMutations from '../hooks/useDuplicatesMutations';
 import { useMessages } from 'core/i18n';
 import useNotDuplicates from '../hooks/useNotDuplicates';
 import { useNumericRouteParams } from 'core/hooks';
-import { ZetkinPerson } from 'utils/types/zetkin';
-import ZUIPersonAvatar from 'zui/ZUIPersonAvatar';
-import ZUIPersonHoverCard from 'zui/ZUIPersonHoverCard';
 
 interface NotDuplicatesModalListProps {
   duplicate: PotentialDuplicate;
@@ -24,88 +22,16 @@ const NotDuplicatesModalList: FC<NotDuplicatesModalListProps> = ({
   const { addDuplicatePerson } = useDuplicatesMutations(orgId);
 
   const notDuplicates = useNotDuplicates();
+  const rows = notDuplicates;
+  const isPeopleToMergeTable = false;
 
-  const rows: ZetkinPerson[] = notDuplicates || [];
-
-  const columns: GridColDef[] = [
-    {
-      align: 'left',
-      disableColumnMenu: true,
-      field: 'Name',
-      headerName: messages.modal.possibleDuplicatesColumns.name(),
-      minWidth: 250,
-      renderCell: (params) => (
-        <ZUIPersonHoverCard personId={params.row.id}>
-          <ZUIPersonAvatar orgId={orgId} personId={params.row.id} />
-          <Typography
-            sx={{ alignItems: 'center', display: 'inline-flex', marginLeft: 2 }}
-          >
-            {params.row.first_name + ' ' + params.row.last_name}
-          </Typography>
-        </ZUIPersonHoverCard>
-      ),
-      sortable: false,
-    },
-    {
-      align: 'left',
-      disableColumnMenu: true,
-      field: 'E-Mail',
-      headerName: messages.modal.possibleDuplicatesColumns.email(),
-      minWidth: 250,
-      renderCell: (params) => (
-        <Typography
-          sx={{ alignItems: 'center', display: 'inline-flex', marginLeft: 2 }}
-        >
-          {params.row.email}
-        </Typography>
-      ),
-
-      sortable: false,
-    },
-    {
-      align: 'left',
-      disableColumnMenu: true,
-      field: 'phone',
-      headerName: messages.modal.possibleDuplicatesColumns.phone(),
-      minWidth: 250,
-      renderCell: (params) => (
-        <Typography
-          sx={{ alignItems: 'center', display: 'inline-flex', marginLeft: 2 }}
-        >
-          {params.row.phone}
-        </Typography>
-      ),
-      sortable: false,
-    },
-    {
-      align: 'right',
-      disableColumnMenu: true,
-      field: 'button',
-      headerName: '',
-      minWidth: 250,
-      renderCell: (params) => {
-        return (
-          <FormControl
-            sx={{
-              display: 'flex',
-              flexDirection: 'row',
-              gap: 2,
-              justifyContent: 'flex-end',
-            }}
-          >
-            <Button
-              onClick={() => addDuplicatePerson(duplicate.id, params.row)}
-              size="small"
-              variant="outlined"
-            >
-              {messages.modal.isDuplicateButton()}
-            </Button>
-          </FormControl>
-        );
-      },
-      sortable: false,
-    },
-  ];
+  const columns: GridColDef[] = getColumns(
+    addDuplicatePerson,
+    duplicate.id,
+    isPeopleToMergeTable,
+    messages,
+    orgId
+  );
 
   return (
     <Box display="flex" flexDirection="column" overflow="hidden" padding={2}>
